refactor: Remove AI verbosity on comments and tests
