+++ packages/serverpod_database/test/expressions/expression_test.dart
@@ -64,21 +64,14 @@ void main() {
   group('Given a NotExpression when wrapExpression is called', () {
     var notWrappedExpression = NotExpression(const Expression('TRUE'));
 
-    test(
-      'then the resolved argument is used and parenthesized (not the raw inner '
-      'expression). See issue #5294.',
-      () {
-        expect(
-          notWrappedExpression.wrapExpression('x IN (SELECT y FROM z)'),
-          'NOT (x IN (SELECT y FROM z))',
-        );
-      },
-    );
+    test('then the resolved expression is wrapped in NOT.', () {
+      expect(
+        notWrappedExpression.wrapExpression('x IN (SELECT y FROM z)'),
+        'NOT (x IN (SELECT y FROM z))',
+      );
+    });
 
-    test('then toString stays unparenthesized and uses the raw inner '
-        'expression', () {
-      // toString() must remain stable: it is relied on by the pinned tests
-      // and is never used to render a many-relation where clause.
+    test('then toString uses the raw inner expression.', () {
       expect(notWrappedExpression.toString(), 'NOT TRUE');
     });
   });
