+++ tests/serverpod_test_server/test_integration/database_operations/entity_relations/many_to_many/find/where_none_test.dart
@@ -48,6 +48,35 @@ void main() async {
     );
 
     test(
+      'when fetching models filtered by NOT none many relation then result is as expected',
+      () async {
+        var students = await Student.db.insert(session, [
+          Student(name: 'Alex'),
+          Student(name: 'Isak'),
+          Student(name: 'Lisa'),
+        ]);
+        var courses = await Course.db.insert(session, [
+          Course(name: 'Math'),
+          Course(name: 'English'),
+        ]);
+        await Enrollment.db.insert(session, [
+          Enrollment(studentId: students[0].id!, courseId: courses[0].id!),
+          Enrollment(studentId: students[0].id!, courseId: courses[1].id!),
+          Enrollment(studentId: students[1].id!, courseId: courses[1].id!),
+        ]);
+
+        var studentsFetched = await Student.db.find(
+          session,
+          where: (s) => ~s.enrollments.none(),
+        );
+
+        var studentNames = studentsFetched.map((e) => e.name);
+        expect(studentNames, hasLength(2));
+        expect(studentNames, containsAll(['Alex', 'Isak']));
+      },
+    );
+
+    test(
       'when fetching models filtered by filtered none many relation then result is as expected',
       () async {
         var students = await Student.db.insert(session, [
